Add logic to help guard against invalid selections
The Scope section controls what further options are enabled for the
user.

* If an option other than invited individual status is selected, only
the Signing-as-individual option to choose the invited individual status
will be disabled.
* If the invited individual status is selected, it becomes the only
choice in the Signing-as-individual section. Furthermore, all fields
for the Signing-as-entity will be disabled.

Implementation:

The Scope sync function takes new dependencies on elements in the
Signing section. As a result, its script block is moved down below the
definition of those elements. The sync function now updates the
individual and entity signing sections when options change.

Asignment of disabled status and accompaning visual affordance for the
related label is now handled by a helper function. The helper function
provides an additional optional service of setting/clearing an
"enabled exception" status on the input field using a data- attribute.
These exceptions are used when any of the signing details elements are
toggled: the details elements' default sync handler behavior is to
disable all inputs on close and enable all inputs on open; inputs with
the noted exception are forced-disabled when the details element is
toggled open to be consistent with their disabled UI affordance made by
the Scope section's sync handler.