Link to instructions for unverified individuals and entities
It was not clear to me, as a WHATWG spec editor, that verifying a
participant was a step I needed to do. Providing a link here, on the
page which is linked to from the IPR bot failure, makes it clear that
spec editors are empowered and expected to verify individuals and
entities, and links to the existing good instructions on how to do so.