Move to the Node.js built-in test runner
This is the first step towards moving this package to use native JS modules instead of CommonJS. Jest does not support mocking native JS modules, so we move to the built-in Node.js test runner instead.

This requires using Node.js v24, as that's the first version that supports mocking JSON modules. It also requires using --experimental-test-module-mocks, which is a bit unfortunate, but at least we're only using it during testing.